feat: add missing reset_workbench e2e test, e2e test some invalid cases.
Most of the invalid cases are handled by the get_chapters_test, but might as well add some shallow ones to make sure those are being activated.

Additionally, this commit improves the valid_project fixtures to be more useful for diff testing, and also to be a little bit more correct bonsai-wise.

Signed-off-by: Alexander Skvortsov <sasha.skvortsov109@gmail.com>