feat: for stage 2, still take `Position.t list` as input to Apple actions
This is a little less descriptive, but lets us avoid changes to `Apple.t` in stage 3 of the tutorial, which is worth it for the added simplicity.

Ideally we'd pass `Game_elements.t` to `Apple`'s actions, but then we get a circular dependency. Separating out `Apple.t` from `Apple.computation` would also be doable, but would add even more files, further complicating the tutorial.

Signed-off-by: Alexander Skvortsov <sasha.skvortsov109@gmail.com>