MAKE 👏 ILLEGAL 👏 STATES 👏 UNREPRESENTABLE
The game doesn't start until the user clicks on the page, but that restarts everything, so the position shown at first is meaningless.

Therefore, if status is `Not_started`, there should be no snake, score, or apple.

As an alternative, we could separate "Spawn" and "Start" into separate actions, and run a "Spawn" on lifecycle activate, but that would be more complex than this change.

Signed-off-by: Alexander Skvortsov <sasha.skvortsov109@gmail.com>